minor adjustments to formatDistance and count

+++ src/utils/functions.js
@@ -55,7 +55,7 @@ const formatters = {};
 
 export function formatCount(n, lang = "en") {
     if (!formatters[lang])
-        formatters[lang] = new Intl.NumberFormat(lang, { compactDisplay: "long", notation: "compact" });
+        formatters[lang] = new Intl.NumberFormat(lang, { compactDisplay: "short", notation: "compact" });
     let num = n;
     if (typeof n === "string") num = +n;
     return formatters[lang].format(num);
